+++ Line-Segment-Intersection/Line-Segment-Intersection/status_structure.h
@@ -17,7 +17,7 @@ public:
 	status_structure();
 	~status_structure();
 // Member Functions
-	void add(line_segment* lineSegToAdd, bool useMinorStep = false, float position = FLT_MAX);
+	void add(line_segment* lineSegToAdd, bool useMinorStep = false, double position = FLT_MAX);
 	void find(std::vector<line_segment*>& linesL, std::vector<line_segment*>& linesC, event_point& ep);
 	bool findAdjacentSegments(line_segment*& leftSeg, line_segment*& rightSeg, event_point& ep);
 	int findBoundariesOfUnion(line_segment*& leftSeg, line_segment*& leftmostSeg,
